Cleanup: Remove PyQt6, consolidate Qt compatibility, update docs
Major cleanup changes:

1. Remove PyQt6 support - only support PyQt5 and PySide6
   - Removed from pyproject.toml optional dependencies
   - Removed from GitHub Actions workflow matrix
   - Removed PyQt6-specific enum compatibility code

2. Consolidate Qt compatibility layer
   - Moved all Qt mock classes from launcher.py to qt_compat.py
   - Single source of truth for Qt imports and compatibility
   - Simplified launcher.py imports

3. Remove unnecessary files and directories
   - Deleted scripts/generate_test_binary.sh
   - Deleted tests/_resources/ directory and all contents
   - Deleted tests/integration/README.md
   - Removed unittest discovery check from tests/integration/__init__.py

4. Reorganize tests
   - Moved tests/unit/test_anti_deob.py to tests/integration/
   - This test requires idapro module, so it's an integration test

5. Simplify docker-compose.yml
   - Removed idapro-tests and idapro-tests-9.2 services
   - Only keep idapro-91 and idapro-92 development services

6. Update README.md with comprehensive documentation
   - Added installation instructions
   - Added Quick Start section with basic examples
   - Added worker script example
   - Added advanced WorkerLauncher example
   - Updated testing instructions for unit and integration tests
   - Added Docker development environment instructions
   - Removed references to deleted files

All tests pass with PyQt5 and PySide6 only.
