+++ .github/workflows/python.yml
@@ -53,7 +53,7 @@ jobs:
         strategy:
             matrix:
                 python-version: ["3.10", "3.11"]
-                qt-framework: ["pyqt5", "pyqt6", "pyside6"]
+                qt-framework: ["pyqt5", "pyside6"]
         env:
             PYTHONPATH: ${{ github.workspace }}
             QT_QPA_PLATFORM: offscreen
