+++ pyproject.toml
@@ -14,7 +14,6 @@ ci = [
     "pytest-cov>=4.0.0",
 ]
 pyqt5 = ["PyQt5>=5.15.0"]
-pyqt6 = ["PyQt6>=6.0.0"]
 pyside6 = ["PySide6>=6.0.0"]
 
 [project.scripts]
