+++ tests/integration/__init__.py
@@ -1,12 +1,4 @@
-"""Integration tests for IDA Taskr with real IDA Pro.
+"""Integration tests for IDA Taskr with real Qt environments.
 
-These tests require pytest and a real IDA Pro environment.
-They will be skipped if run via unittest discovery.
+These tests require pytest and a Qt framework (PyQt5 or PySide6).
 """
-
-import sys
-
-# Check if we're being imported by unittest - if so, skip this entire module
-if 'unittest' in sys.modules and 'pytest' not in sys.modules:
-    import unittest
-    raise unittest.SkipTest("Integration tests require pytest - use 'pytest tests/integration/' to run them")
